Fix startup_test failures on Windows buildbot.
The problem seems to be an error when overwriting file which is in use
(that's how EvictFileFromSystemCache works on Windows). So I used
a standard workaround for such Windowsic problems and created a wrapper
which retries 10 times. Similar code was there before my porting patch.

TBR=agl

Review URL: http://codereview.chromium.org/40160

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10981 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e9cde1473cfbc973f3cc856067d1ca5f5bd99a86