POSIX: Rewrite IPC's interaction with FileDescriptor
The FileDescriptor API is clearly too hard to use. It's the only IPC
data type which is non-POD and serialising an invalid file descriptor
is fatal to Chrome on POSIX. The use of Maybe is possibly non-obvious
to non-functional programmers.

This patch merges Maybe and FileDescriptor so that serialising invalid
file descriptors is permitted and results in -1 at the other end.
(Serialising /closed/ a file descriptor is still fatal.) Also, it adds
a pointer in base/file_descriptor.h to instructions for its use with
IPC. Although it's generally bad practice to mention IPC in base, in
this case I cannot find another suitable location.

Review URL: http://codereview.chromium.org/39208


git-svn-id: http://src.chromium.org/svn/trunk/src@11041 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b2b2bfc8eb71a0c43f5d96597b68f7261b11274d