More WebKit API action (chromium side).
Adds the following methods to WebKitClient:

  decrementStatsTable
  incrementStatsTable
  traceEventBegin
  traceEventEnd

The implementation of initV8CounterFunction is now a no-op.  V8Proxy.cpp no
longer calls it, and I moved the corresponding work out of WebKit.  The
embedder now calls the appropriate V8 API before initializing WebKit.

Includes some cleanup to WebString and WebCString.  I decided that I prefer
"data" over "characters" since the latter may be confusing in some cases.  For
example, the elements of a WebString or WebCString may not be individual
characters but may be part of a character.  "data" is also consistent with
basic_string<T> and so should be plenty familiar.

There is wanton disregard for the Chrome+JSC build in this CL, but that's
OK.  We have already crossed that bridge.

R=dglazkov

Review URL: http://codereview.chromium.org/39288

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11254 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a3c7895837d2da6e2398bda96532acb38ad8d541