Ensure that a listener's OnChannelConnected is called before OnMessageReceived.
A race condition existed since the listener thread could dispatch messages in two ways: a task being posted to it, or because the dispatch messages event is set.  The latter is needed to avoid deadlock when multiple SyncChannels are used on the same listener thread.  If the latter method was used to dispatch a message, it might occur before the task that calls OnChannelConnected gets run.

The fix is to call OnChannelConnected manually if we notice a message is being dispatched before it's called.

Note: I couldn't think of a way to test this since it depends on a race condition in the listener message looop (whether the task or objectwatcher gets run first).

Review URL: http://codereview.chromium.org/42113

git-svn-id: http://src.chromium.org/svn/trunk/src@11522 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7ef421405860641ca06fa1361e732543761c0b3e