RSS feed support (part 1)
Part 1 is RSS feed auto-discovery. 

This will parse the web page header to find the feeds in the document and notify
the browser to display the RSS icon in the toolbar. You can click on the icon,
but it will just navigate to the first feed on the page, which (unless it has
been designed to be browser friendly) will just dump XML as text on the user.

For this reason I have disabled the code that makes the RSS icon appear and
intend to enable it when we have a good landing page to display the XML.

Review URL: http://codereview.chromium.org/43109

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11640 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b1fa19a7ca4c8a46c189469b672787cb4877e12b