Basic windowless plugins, try 2.
In response to Dean's comment on http://codereview.chromium.org/39105,
I redid that patch to put the X munging into the plugin implementation.  This
won't work for multiproc, but it's near the correct place and many things will
need to be changed for multiproc.

Review URL: http://codereview.chromium.org/42056

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11674 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 43a025f3b87906780d0edf02d5fe0acb62c51459