Basic windowless plugins, try 2.
In response to Dean's comment on http://codereview.chromium.org/39105,
I redid that patch to put the X munging into the plugin implementation.  This
won't work for multiproc, but it's near the correct place and many things will
need to be changed for multiproc.

Review URL: http://codereview.chromium.org/42056

git-svn-id: http://src.chromium.org/svn/trunk/src@11674 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 570309992a8e1da83ed069fc5cb8db023a9f187c