Fix 2 layout tests under editing\selection.
These tests test the Mac behavior of selecting words on
right-click. We don't do this on Windows so the caret
should be set instead of a selection being formed.

Interestingly, one test changed upstream to not
select on right-click so we match that except for
font metrics.

Review URL: http://codereview.chromium.org/42256

git-svn-id: http://src.chromium.org/svn/trunk/src@11786 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 62ef33a398badc27fc02be535d4149dc4de5e346