SSLPolicy Fix: Step 6.
Merge in changes to SSLHostState.  We now can store whether a specific origin is "broken," which is the key new bit of state that we need to share between tabs.

Currently, there is a naming inconsistency between the SSLManager names and the SSLHostState names.  I'll clear this up when I merge in the new SSLManager.

R=jcampan
BUG=8706

Review URL: http://codereview.chromium.org/42274

git-svn-id: http://src.chromium.org/svn/trunk/src@11891 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7f5da0095732650ff81bb054b46f9a4432d247ea