SSLPolicy Fix: Step 6.
Merge in changes to SSLHostState.  We now can store whether a specific origin is "broken," which is the key new bit of state that we need to share between tabs.

Currently, there is a naming inconsistency between the SSLManager names and the SSLHostState names.  I'll clear this up when I merge in the new SSLManager.

R=jcampan
BUG=8706

Review URL: http://codereview.chromium.org/42274

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11891 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a0d17d67fefa86acfd5e57885f8ebd39b5dd5882