Use anonymous shared memory for VisitedLinks.
We already pass a handle to the shared memory via IPC but, when we
create it, we create a named segment. These segments are not garbage
collected on POSIX.

Since nobody actually uses the name of the segment, use an anonymous
segment instead and remove the functions.

Review URL: http://codereview.chromium.org/48087


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11900 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 587b048281de6a848b4db9b161e763a8a0d384d5