Fixing a few minor issues found during static analysis of the Chromium codebase.
find_bar_controller.cc:

Unused local variable created.

hwnd_view.cc:

The local definition of parent overshadows param passed in.

google_update.cc:

Uninitialized member variable.

webframe_impl.cc:

The if was overly constrained. At the top of the function there is an if (reset) { ... return; } so we don't need to check for !reset in the next if statement. This functionality is covered by a ui_test, which passes with this change.

Review URL: http://codereview.chromium.org/42321

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11964 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 01764268d1cf065c61d1fd3580c55916d8585421