ScopedVariant implementation.
A class for automatically freeing a COM VARIANT at the
end of a scope.  Additionally provides a few functions to
make the encapsulated VARIANT easier to use.
Instead of inheriting from VARIANT, I took the containment
approach in order to have more control over the usage of the
variant and guard against memory leaks.


Review URL: http://codereview.chromium.org/46059

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12081 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6b8588b17225baff145317538e65908ac356b1a2