The street view in Google maps, which is implemented with a windowed flash plugin would draw over the iframe DOM element, effectively hiding it. This is handled by our IFrame shim geometry calculation, where we subtract the rect of any IFrame above the plugin in the ZOrder from the plugin rect.
Webkit calls Widget::setFrameRect at various times, during layout/size changes/paints, etc. The reason this bug
showed up, was due to an optimization in our setFrameRect implementation, where we would bail out if the rect 
passed in was the same size as the current plugin rect. Basically the IFrame appears above the plugin in the ZOrder
much later, which causes us to return without sending over the cutout rects to the browser when it moves the plugin
windows.

Fix is to move the rects equality check to WebPluginImpl::setFrameRect, where we don't send over the UpdateGeometry IPC to the plugin process if the rects are the same.

WebPluginImpl used to implement the webkit Widget interface a long time ago. This is no longer the case. So some
of the functions don't need to be virtual anymore. I also made this change.

This fixes bug http://b/issue?id=1722236 and http://code.google.com/p/chromium/issues/detail?id=8858

Bug=1722236,8858


Review URL: http://codereview.chromium.org/42413

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12179 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b6ddbc67ffd7ff46bdd80e6b51869e54c21effdf