Fixed the following layout test:     LayoutTests/http/tests/xmlhttprequest/redirect-cross-origin-post.html
The crash happens in simple_resource_loader_bridge.cc, there is a
case when MaybeUpdateUploadProgress is posted as a task and in the
queue while a redirect response by the server cancels the upload.
Leaving request_->get_upload() returns NULL, and of cource using of
the pointer failes.

Review URL: http://codereview.chromium.org/50077

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12379 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a7bb32f998a163c692b5907b73e958f9527cee95