Fixed the following layout test:     LayoutTests/http/tests/xmlhttprequest/redirect-cross-origin-post.html
The crash happens in simple_resource_loader_bridge.cc, there is a
case when MaybeUpdateUploadProgress is posted as a task and in the
queue while a redirect response by the server cancels the upload.
Leaving request_->get_upload() returns NULL, and of cource using of
the pointer failes.

Review URL: http://codereview.chromium.org/50077

git-svn-id: http://src.chromium.org/svn/trunk/src@12379 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f7a3b6fbf2ede0ec40a540c8712ecbbb2eda56d9