Fix data url security tests.
The problem was that the security token in V8 was set to the string
"null" for all data urls.  Therefore, the access check succeeded.  The
cause of the regression was that securityOrigin's toString
implementation was changed in WebKit to return "null" instead of
String().

R=kasperl@chromium.org
TBR=scherkus@chromium.org
Review URL: http://codereview.chromium.org/53049

git-svn-id: http://src.chromium.org/svn/trunk/src@12444 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6ae99fe4be6987baaa64bd53ef1f565977f19c6c