Fix fast shutdown
The check for unload handlers was backwards.  This bug was introduced in r7901 <http://codereview.chromium.org/17453>.

Also, remove the call to process_.Close().  When we Terminate the process, we get a IPC channel error and BrowserRenderProcessHost::OnChannelError gets called.  This method needs the handle so it can call DidProcessCrash.  It also does the call to process_.Close() for us.

Review URL: http://codereview.chromium.org/42653

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12597 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c16b6892f3ab32d687c81d79800f6fcb971be60d