This CL fix issue 5724 - RTL: Windows context menu with Alt+Space Bar should be displayed on right side of the window. (http://crbug.com/5724)
and issue 5738 - RTL: App drop-down menu should be displayed below the favicon
(http://crbug.com/5738)

In RTL UI, the x-axis of the system menu was computed as the MirroredXCoordinateInsideView(x-axis in LTR UI);
Also, the menu was right-aligned relative to the x-axis.

It fixes the system menu position in the following 3 situations:
1. press alt+space (such as in application short-cut, in chrome browser, in pop-up dialog such as option/clear browsing data), the system menu appears on the right side of the window (instead of left side, without the fix).

2. click favicon of application shortcut, the system menu appears on the right side of the window (instead of left side, without the fix).

3. right click on chrome browser tab, system menu appears at
left side of the click (instead of right side of the click).

Please be noted that:
System menu itself is not yet mirrored. This CL only fixes its display position. 
Review URL: http://codereview.chromium.org/53120

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12631 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b17f97114d88f0e35d0fa6b9c527f05f24c592ef