Disk Cache: Second pass (and final) to allow multiple instances of BackendImpl.
This cl takes care of all the histograms on the disk cache. Most
of them have to be splitted in three so that we get separate data
from different cache types. There are a few places where the complexity
of splitting the histogram is not worth it so we just keep either all
data together (if it makes sense), or just ignore data for some types of
caches.

note: Having multiple versions of a histogram but only one "active" for a
given client is not the same as having multiple histograms working at the
same time for different objects.

Review URL: http://codereview.chromium.org/42682

git-svn-id: http://src.chromium.org/svn/trunk/src@12692 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 130332edbf1a96b82fdcf81435e49ded6d30d585