Reland changes to fix the problem that DOM object is deleted in the GC thread, not owning thread.
This patch contains the fix to the Purify failure: memory leaks of thread specific data for main thread. The fix is to switch to using the static maps for main thread.

Compared with previous patch being reviewed, the main change in this patch is in V8DOMMap.cpp: I introduce NonMainThreadSpecificDOMData and MainThreadSpecificDOMData that are both derived from ThreadSpecificDOMData.


Review URL: http://codereview.chromium.org/49044

git-svn-id: http://src.chromium.org/svn/trunk/src@12757 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c0854b820d63e7cd72d6df9083a8049d01dce133