In certain cases, the coordinates used for pattern rendering can go negative, eg., when a negative translation is applied in the shader matrix.  This causes the rasterizer to blow up, since it accesses memory outside the pattern bitmap.
Since the integer modulus operator for C++ has unspecified behaviour with
negative arguments, its value may go negative.  In this case, we must offset
by the modulus to make it positive again.  I decided to do this at a low level
in skia, since it seems better to make skia robust than to pray that you get
strictly non-negative translations.

(There is a fix for this in src/skia/tile_patch.diff, but it's wrong.)

This patch should be upstreamed to skia.

Review URL: http://codereview.chromium.org/55044

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12789 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6ae27373a49f177822d8f7de5e64f03ec293c162