The layoutTestController had its queueScript method broken into two: queueLoadingScript and queueNonLoadingScript.  The behavior of the work queue also changed to be a bit simpler.  Now, each WorkItem reports whether or not it kicked off a load.
There is a bit more work to do here to determine if work items actually
resulted in a load.  For now, I am just guessing, which is almost always
correct.  I will follow-up with the other part of the CL once this one clears
the bots.

This change corresponds roughly to:
http://trac.webkit.org/changeset/42082

BUG=9581
R=dglazkov

Review URL: http://codereview.chromium.org/56164

git-svn-id: http://src.chromium.org/svn/trunk/src@12991 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6a1bd85d3b2610b7e36b9849a868731a4bc9a951