Fix tests that depend on setCustomPolicyDelegate and waitForPolicyDelegate.
Changes:

1-  Need to support a second parameter to setCustomPolicyDelegate that
controls whether subsequent loads are permitted or not.  Previously,
we would always deny any subsequent loads.

2-  Need to change the string that we produce when writing out the URL
that is impacted by a custom policy delegate.  We need to write out only
the filename part of the URL.  (We need to do this only for file:// URLs.)

3-  Add LayoutTestController::waitForPolicyDelegate and corresponding
method on TestWebViewDelegate.

4-  Improve the way we reset the TestWebViewDelegate.  Now, instead of
having support for reseting individual fields, we just reallocate the
delegate.  This requires adding a SetDelegate method to WebView.

BUG=9582
R=dglazkov

Review URL: http://codereview.chromium.org/56169

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13014 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4ffed73d40f303949f3c40db5d200f39d07976ac