A workaround fix for Issue 9186.
This crash is caused in a unnecessary check whether or not to add a "text direction" submenu to a context menu. Since we don't have a "text direction" submenu and this check is unnecessary now, this change set a page setting that disables this check.

BUG=9186
Review URL: http://codereview.chromium.org/42690

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13072 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a19a6132b599e36ee88219df8eeab664694e2ca8