A workaround fix for Issue 9186.
This crash is caused in a unnecessary check whether or not to add a "text direction" submenu to a context menu. Since we don't have a "text direction" submenu and this check is unnecessary now, this change set a page setting that disables this check.

BUG=9186
Review URL: http://codereview.chromium.org/42690

git-svn-id: http://src.chromium.org/svn/trunk/src@13072 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 870320dbfc5b70c745b5204340466762203deeeb