Remove some uses of custom tab contents types from some tests.
This required redoing the tests that used it to use the new RenderViewHost
test harness. This is mostly straightforward, but required more work for the
NavigationController tests.

I removed a few random places that were testing for the tab types. I also
made different versions of a few functions that don't take tab contents types
next to the existing ones that do, so my touched code won't have to be changed
again when I remove the tab contents types.

Review URL: http://codereview.chromium.org/62039

git-svn-id: http://src.chromium.org/svn/trunk/src@13132 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 88e31fd85796789190af6b5330e4745cbc295f28