Fix a problem with DrawRectInt at the edges of a ChromeCanvas.
It seems there used to be an old bug in Skia, and we explicitly set a stroke
width of 1 instead of 0 to work around it.  Internally in Skia this goes down
a more complex path, creating a path from the rect, and then filling the path.
This was causing problems at the edges of the canvas, for example if you were
to DrawRectInt(0, 0, width - 1, height - 1).

Switch some SkRects to SkIRects.  Internally Skia will still convert them to scalars, but it makes our code a bit cleaner.

Review URL: http://codereview.chromium.org/62099


git-svn-id: http://src.chromium.org/svn/trunk/src@13254 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1c6479db263a6ccce84ba9bf008ef7b27cf1a717