Remove some unused stuff from basictypes.h.
My original goal was just to remove down_cast, since it uses RTTI, and I am
going to update the build with -fno-rtti.  The other stuff I just noticed is
unrelated and unused in our code, so we might as well remove it, since it
hasn't proven to be useful.

Review URL: http://codereview.chromium.org/63121


git-svn-id: http://src.chromium.org/svn/trunk/src@13349 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 42e9695d67f2d2a75e414aa5f9d52af1ff567bbd