Let WebKit draw the active highlight for FindInPage, instead of using the SelectionController.
This fixes a couple of bugs, such as: 

Active match highlighting disappears when mouse selection
is made within the page and selection color is orange (6145).

and 

Selection color becomes orange in some cases on tabs that
don't have a Find box open.

and

Gmail jumping to the first match if on FindNext if Find box
is closed and reopened (9795).

NOTE: This patch depends on my patch to WebKit that just
got accepted. See:
https://bugs.webkit.org/show_bug.cgi?id=25102

BUG=6145, 9795
TEST=Find something on a page that results in multiple
matches being found. Highlight another word that is not
currently highlighted. You should see both active and 
inactive highlights (orange and yellow) and a blue colored
selection highlight.

Review URL: http://codereview.chromium.org/66016

git-svn-id: http://src.chromium.org/svn/trunk/src@13510 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1946c96d66578645e9048d6b47e1665a46e238e4