Make HTMLSelectElement and HTMLOptionsCollection return collections of options elements if there are multiple options that share a name.
Currently Safari and chromium differ in their behavior in this area. Safari returns
multiple options in a collection if necessary, but only for HTMLOptionsCollection. For 
HTMLSelectElement it always returns null.
Chromium returns a single option no matter what, but for both HTMLSelectElement
and HTMLOptionsCollection. It's also very slow at this, as it bypasses a webkit cache.

The HTML5 doc states that a collection should be returned if necessary for both.
The relevant parts of the spec are:
http://dev.w3.org/html5/spec/Overview.html#the-select-element
and
http://dev.w3.org/html5/spec/Overview.html#htmloptionscollection-0

So if you have:

<select id="sl1"><option value="Value" name="test" /></select>

and in JS code you do:
var test = document.getElementById("sl1").test

test should be a single HTMLOptionElement.

If instead you have:

<select id="sl2"><option value="Value1" name="test" /><option value="Value2"
name="test" /></select>

test should be a collection with 2 HTMLOptionElements. 

This bug: https://bugs.webkit.org/show_bug.cgi?id=25191 has been filed against
webkit, with a new layout test to confirm the behavior

Review URL: http://codereview.chromium.org/67148

git-svn-id: http://src.chromium.org/svn/trunk/src@13760 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c7b659e5b529f40052cb952feda02deeb6a8b9b5