Linux/Skia: remove VDMX and Harfbuzz
In order to unfork our Skia, VDMX parsing has been moved into WebKit
in r42548.

The Harfbuzz stuff should have been functional, but noone on the
WebKit side wants to review a bidi patch so that side never actually
landed. At this point it's probably better to remove the current
Harfbuzz support in Skia with a view to jumping straight to the
version agreed with Mike Reed at some point in the future.

http://codereview.chromium.org/75011


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13795 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 57285b5b68cc5b9340d5ff2e48ec95ce3226c187