Rebaseline a bunch of tests after the merge.
In http://trac.webkit.org/changeset/41683 , we accidentally
broke the skia drawRect method.  E.g., images that fail to load
didn't get an outline.  This was landed in a merge and the
tests rebaselined to not have the rectangle (there were other
skia changes so this was missed).  In
http://trac.webkit.org/changeset/42547 , we restored the rectangle,
so this is me rebaselining on windows.

I manually verified that these only change borders.  I spot-checked
a few upstream mac results and this seems to match.

Linux is coming next.

Review URL: http://codereview.chromium.org/75024

git-svn-id: http://src.chromium.org/svn/trunk/src@13815 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: bdd3acf149d1d1598a3e1c371399d5cc2e20509a