Rebaseline some tests on linux.
inlineBoxWrapperNullCheck.html is a test that changed due to drawRect.

generated.html is the same output as windows, just with anti-aliased
fonts on linux.

The other three (bug56201 bug23847 008) are tests that have checksums
checked in, but no png.  I'm checking in the pngs so they're there
for comparison.

Review URL: http://codereview.chromium.org/75025


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13849 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 80b3dac919903f4df8c67d3d204a7fe36dea4345