Fix 10573: Dismissing Find-in page doesn't set focus to the link found.
We no longer use the selection controller to
highlight the active match. Before this change,
the focus would not be set if the user had changed
the selection. After this change, the focus will
be set unless the user has selected something on
the page.

I also wrote an in-browser unit test for this to
catch this regression in the future, but it is
disabled due to problem with running multiple
in-process browser tests in a row (teardown
problem).

BUG=10573
TEST=Covered by in process browser test now, see
bug for repro steps.

Review URL: http://codereview.chromium.org/79024

git-svn-id: http://src.chromium.org/svn/trunk/src@13945 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 71002d6e6785108c4bbe261893724077a5d159a9