Refactor DialogDelegate so the DialogButton enum is in cross platform code.
This is mostly a find & replace in our code to use the new location
of the enum.

I also deleted some GetDialogButtons methods in child classes
because they were identical to the parent version in DialogDelegate.

This will allow more code to be enabled on linux/mac in the
automation provider.

Review URL: http://codereview.chromium.org/88008

git-svn-id: http://src.chromium.org/svn/trunk/src@14146 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4cbe3319260fa2cf6c25329afbd116ab1a34b03d