Refactor DialogDelegate so the DialogButton enum is in cross platform code.
This is mostly a find & replace in our code to use the new location
of the enum.

I also deleted some GetDialogButtons methods in child classes
because they were identical to the parent version in DialogDelegate.

This will allow more code to be enabled on linux/mac in the
automation provider.

Review URL: http://codereview.chromium.org/88008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@14146 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4689063df84f7437f64ef5c2327f04e73e0ae4c2