Always enable copy-link-address in context menu.
Added a new field into ContextMenuParam only for this purpose. The new field
won't be validated in frontend processes. In this way, even if renderer
processes are going to mad, the frontend would be OK if the frontend uses this
new field only for copying into the clipboard.

Fix checked in for Shinichiro Hamaji
Original code review: http://codereview.chromium.org/91002

BUG=2725

git-svn-id: http://src.chromium.org/svn/trunk/src@14256 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 19bb3a2751e0343f5659618925c5287f11cbb2e2