Always enable copy-link-address in context menu.
Added a new field into ContextMenuParam only for this purpose. The new field
won't be validated in frontend processes. In this way, even if renderer
processes are going to mad, the frontend would be OK if the frontend uses this
new field only for copying into the clipboard.

Fix checked in for Shinichiro Hamaji
Original code review: http://codereview.chromium.org/91002

BUG=2725

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@14256 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4168fbaf7b5967e74d3453952111033c16463cc0