Fix DCHECK when typing certain kinds of input on certain profiles.
This is actually a very old bug that some of my recent changes exposed more widely.  In PromoteOrCreateShorterSuggestion(), we could create a totally bogus search base, which, if added to the history matches, would then break.  In the old code, we'd create this match, but usually not add it to the history match set because we'd already have a What You Typed match; in my new code that's frequently no longer the case.

BUG=10926
Review URL: http://codereview.chromium.org/92154

git-svn-id: http://src.chromium.org/svn/trunk/src@14470 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e596325f12cced2c315bf90cb1edd49516377c0f