Massively improve the performance of dragging a window over our window.
Our toolbar background is a 1 pixel wide image, which we tile by drawing a
pixbuf over and over.  The way the toolbar is laid out, everything shares
the same window.  The toolbar is painted first, then the buttons on top of
it, etc.  If the toolbar painting is slow, you can see gray trails when you
drag a window over and the exposes are slow to come back to X.

Honor the damage rect, so when we get tons of tiny exposes we aren't doing
this really inefficient tiling over the entire toolbar.  Improving the tiling
performance needs to be done separately.

Review URL: http://codereview.chromium.org/100036


git-svn-id: http://src.chromium.org/svn/trunk/src@14604 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f79954243c4b85530769c445131d824fce238141