Make the font scale and the font family to use for UI 'OS-dependent' (Win XP vs Win Vista/ Win 7).
In r13773, the localizable font size/family override was added to work around the tiny
font issue for some Indian languages. Initially, I thought the 'tiny UI font' issue was completely taken care of on Vista/Win 7 (it's resolved for Malayalam) and added a check for that.

It turned out that Indian fonts on Vista and Windows 7 were updated, but the issue has not been resolved. As a result, he scale and the fonts for Windows XP do not work on Vista/Win 7 and we need to use different values for Vista/Win 7. 

Note that what this CL does (having separate resource entries in a single reosource bundle for Win XP and Vista/Win 7 and picking one at run-time) is a temporary work-around for the branch and for the trunk, we need to have a better way of handling platform-dependent resources. It's not just for different flavors of Windows but also for taking care of differences between Win, Mac and Linux.
 
BUG=7319
TEST=On Vista/Win7/Win XP, run Chrome with '--lang=XX' where XX is hi, mr, kn, gu, te, ta, bn, ml (and or on Vista/Win 7) and make sure that UI font size looks ok and there's no truncated strings/dialog box. 



Review URL: http://codereview.chromium.org/92135

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@14736 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 626513fbae2167b6e6f7407b75d2c166b0c3baa4