A better fix for http://www.crbug.com/2044:  crash on large <canvas> elements.  We disable the __debugbreak only when skia tells us it is prepared to correctly handle a failed (NULL) malloc().  It does this by calling sk_malloc_flags() without SK_MALLOC_THROW.
Note that, since the switch to tcmalloc, the new_handler
was not getting called at all (since tcmalloc doesn't 
support it yet), so this crash is currently unreproducible
in trunk.  In order to test this change, I reverted the 
tcmalloc change in my client.  This is not the case in the 
stable branch, since it doesn't use tcmalloc, so this change 
is still needed there.  (It will also be needed in trunk 
again once mbelshe's re-implementation of the new_handler
is in).

BUG=http://www.crbug.com/2044
Review URL: http://codereview.chromium.org/100163

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@14891 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e0e61f4a7ceae37cf56023c434539bb9a028c102