Added audio/mpeg and video/mpeg to the list of mime types supported by the Activex shim for media player along with their associated extensions.
This fixes http://code.google.com/p/chromium/issues/detail?id=11054, which occurs with embed tags with a source
URL and no mime type. In this case the URL extension is mp3, which is not claimed by any plugin although Quicktime
and media player do handle it. It works in Firefox because the plugin instantiation code in FF correctly gets the
mime type from the http stream and instantiates the plugin accordingly. Webkit attempts to instantiate the plugin
from the embed tag with the source URL. This fails silently in Chrome and Safari displays an error message.

With this fix the only user visible difference would be quicktime handling the mp3 in FF and IE, and media player
handling it in Chrome.

I also moved the media player shim to the end of the plugin list to ensure it does not take over the audio/mpeg and video/mpeg mime types if there is a a plugin which can handle it.

Added plugin tests for the mp3 and mpe extensions.

Bug=11054


Review URL: http://codereview.chromium.org/100180

git-svn-id: http://src.chromium.org/svn/trunk/src@15008 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 52118111369f8200b85d6a55eb60ab0ef77aa714