POSIX: Add a macro for handling EINTR.
On POSIX systems, system calls can be interrupted by signals. In this
case, they'll return EINTR, indicating that the system call needs to
be restarted.

(The situation is a little more complicated than this with SA_RESTART,
but you can read man 7 signal if you like.)

The short of it is that you need to catch EINTR and restart the call
for these system calls:

 * read, readv, write, writev, ioctl
 * open() when dealing with a fifo
 * wait*
 * Anything socket based (send*, recv*, connect, accept etc)
 * flock and lock control with fcntl
 * mq_ functions which can block
 * futex
 * sem_wait (and timed wait)
 * pause, sigsuspend, sigtimedwait, sigwaitinfo
 * poll, epoll_wait, select and 'p' versions of the same
 * msgrcv, msgsnd, semop, semtimedop
 * close (although, on Linux, EINTR won't happen here)
 * any sleep functions (careful, you need to handle this are restart
   with different arguments)

We've been a little sloppy with this until now. This patch adds a
macro for dealing with this and corrects every case of these system
calls (that I found).

The macro is HANDLE_EINTR in base/eintr_wrapper.h. It's safe to
include on Windows and is a no-op there.

On POSIX, it uses GCC magic to return the correct type based on the
expression and restarts the system call if it throws EINTR.

And you can use it like:
 HANDLE_EINTR(close(fd));

Or:
 ssize_t bytes_read = HANDLE_EINTR(read(fd, buffer, len));

*BEWARE* that it will evaluate the argument multiple times, so this is
not safe:
 HANDLE_EINTR(close(FireMissiles()));

http://groups.google.com/group/chromium-dev/browse_thread/thread/41a35b2a457d73a0
http://codereview.chromium.org/100225


git-svn-id: http://src.chromium.org/svn/trunk/src@15102 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c2a68720be7367234e68497840b3712021a4d23b