Defer window.close(), window.resizeTo(), and window.moveTo() actions by posting a task back to the message loop before notifying the RenderWidgetHost to perform these operations. Otherwise, the JS code races with the browser to use the modified window.
BUG=http://crbug.com/6377
BUG=http://crbug.com/6192

Also, because window resizing is asynchronous between the renderer
and browser, added a renderer-side cache of "pending window rect"
to ChromeClientImpl.  Before the change, calling setWindowRect() would
schedule a call to the RenderViewHost to do the work, but a subsequent
call to windowRect() would return the current size, as through the
call to setWindowRect() had never happened.  We cache a pending size
and then schedule a task on the message loop to clear the cached size.
This allows javascript which writes and reads the WindowRect sizes to
work, and once we come out of JS and return to the message loop, we'll 
go back to asking the RenderViewHost for the true values.

This pending_window_size is my only concern about this patch.  It does
pass all layout tests, and the ChromeClientImpl code is executed in the
test_shell.exe code path.

BUG=http://crbug.com/835

TEST=LayoutTests have tests which cover all of these actions.  The
problem is that we run layout tests using the test_shell, not Chrome.

Review URL: http://codereview.chromium.org/101019

git-svn-id: http://src.chromium.org/svn/trunk/src@15214 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 853ebb5cc4baaceabbd5dd8547b21f1c339ffb3b