Lands the fix for bug 8088 from m0 (http://codereview.chromium.org/100022). Here's his description:
Mouse hover out of context menu should un-select menu item.

All Context Menu's which are implemented with ChromeMenu is not properly
un-selecting the menu item when out of focus. The areas where this is affected
are: * Tabs * Bookmark bar * Bookmark managerThe others uses winapi for context
menus. Doesn't use ChromeMenu, uses HMENU, that explanes why the others work.

For this fix, I had to make sure that the MenuPart is always NONE when the mouse
is moved outside of the context menu regions, and that is always the case.
Hence, we can assure to remove selection. The same way it has been done for
SetSelection, I have used the same technique for this, but simpler.

BUG=8088
TEST=For the above mentioned affected areas, I brought up the context menu by
right clicking, and hovered over a menu item, then move the mouse horizontally
outside the context menu region, it deselected the item.

Review URL: http://codereview.chromium.org/102025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@15220 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 37f9c43c02dccb111941582a8def7b9632c9828e