Port the folder selector portion of the BookmarkEditor to GTK.
Mirrors the BookmarkEditorView method, where the contents of BookmarkModel
are copied to a temporary model so changes can be discarded if the user hits
Cancel. In the GTK version, we copy not into another BookmarkModel, but into
a GtkTreeStore, which serves as a model to the GtkTreeView on screen.

Also ports the unit tests.

http://crbug.com/11250

Review URL: http://codereview.chromium.org/99361

git-svn-id: http://src.chromium.org/svn/trunk/src@15257 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b2a9620bfdf92ceb21d49e911d9a8738e29ce4be