Convert autocomplete to use a custom matrix and button cell.
On the user-visible side this change enables mouse hover (the item
under the mouse is highlighted), and lays out the text omnibox v2
style.  Only hover really needed the control change, but I have
another change queued up to style the popup text.  It uses different
colors and shades, and really looked horrible with the NSTableView
highlighting.  [Which is why this part of the CL is coming first.]

TEST=Omnibox popup works and shows highlighting under the mouse in addition to the keyboard-selected item.

Review URL: http://codereview.chromium.org/99310

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@15321 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 44653dcc0d5ae2aa42e006c2bada4fd978fe9ab4