Defer window.close(), resizeTo() and moveTo() actions by posting a task back to the message loop before notifying the RenderWidgetHost to perform these operations. Otherwise the JS code races with the browser to use the modified window.
BUG=http://crbug.com/6377
BUG=http://crbug.com/6192

Cache a pending_window_rect on the render_view (moved from
prior CL where I had it on the chrome_client_impl).  This 
is a short lived cache, and not a complete solution.  It
fixes this case, where a JS script makes multiple operations
and expects the GetWindowSize() to be correct immedately
after having called SetWindowSize().  

BUG=http://crbug.com/835



Review URL: http://codereview.chromium.org/115030

git-svn-id: http://src.chromium.org/svn/trunk/src@15698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 10a9cb693961a8404994ec6d793c1718f4ce755d