Style the omnibox results for v2.
The style is based on the gtk code (autocomplete_popup_view_gtk.cc), with some local tweaks.  I broke this off from the rest of the review so I could experiment with making it nail omnibox-v2, but eventually realized that I don't have enough data for that.  This at least makes things look nicer, and I believe it provides all of the function that will be needed (the remainder is precisely defining fonts and colors and possibly minor highlighting variations).

Summary: I want something concrete so that complaints about v2 wrongness can be specific.

Review URL: http://codereview.chromium.org/99312

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@15777 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 13ec8d5e812caa6c65c3608c8f5c3ff569c00fdf