Fix regression where exact-match URLs could fail to be promoted when longer URLs were typed more.
We should alwys look up whether we have an exact match, regardless of the input type, and use it if we have it.  My change to not add the "exact match" entry for UNKNOWN input should have only affected the case where that match doesn't, in fact, exist.

BUG=1844093
Review URL: http://codereview.chromium.org/113223

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@15786 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f6998cffb2987ba0e32a26ebfb176e18e6ebdffa