Split V8Proxy::retrieveActiveFrame() into two methods.
We now have RetrieveFrameForCurrentContext() and
RetrieveFrameForEnteredContext().

These terms means the same thing they do in V8::Context --
'current' is the top of the js stack and 'entered' is the
bottom.

I needed 'entered' to fix a bug in extensions where if you
call an extension API through the web inspector we get
confused and think the web inspector's view is the one who
called.

Review URL: http://codereview.chromium.org/113085

git-svn-id: http://src.chromium.org/svn/trunk/src@15828 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d0affa8eeea7c781eb285bd4e74fd975994e0914