Re-do r15244 again.
Originally reviewed at http://codereview.chromium.org/100353

Eliminate webkit/glue/webhistoryitem* in favor of adding a
NavigateBackForwardSoon method WebViewDelegate.  This moves
all of the hacky details of how we intercept "history.{back,
forward,go}" into the webkit layer.  My eventual plan is to
teach WebCore how to make this not hacky.

In this version of the CL, TestWebViewDelegate performs the
back/forward navigation directly in NavigateBackForwardSoon
instead of using PostTask to delay it.  I'm doing this to
minimize regressions so that I can hopefully get the rest of
this CL landed.

I also already made the changes to WebKit to force history.
{back,forward,go} to be processed asynchronously.

BUG=11423
TBR=mpcomplete

Review URL: http://codereview.chromium.org/115288

git-svn-id: http://src.chromium.org/svn/trunk/src@15940 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0037411c709c6ce698da5b2c368c41e75b7e73db