A quick fix for Issue 11683.
I noticed this issue is caused by the IMF_AUTOKEYBOARD option of a RichEdit control, an option that allows a RichEdit control to automatically change the input language with a selection. (An omnibox changes a selection to display a styled text.) A RichEdit control automatically turns on the IMF_AUTOKEYBOARD option when we input an RTL character and start changing the input language with a selection. So, this change checks the IMF_AUTOKEYBOARD option and disable it before decorating a text to prevent it from changing the input language while decorating a text.

BUG=11683 "RTL: IME changes when writing a Hebrew character inside an English string in the omnibox"
TEST=Change the input language to English, type "abcd" in an omnibox, change the input language to Hebrew, type an 'a' key, and verify the input language is Hebrew.
Review URL: http://codereview.chromium.org/113394

git-svn-id: http://src.chromium.org/svn/trunk/src@16154 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 36905124dc87857c63ad3ba2c20449ea10ae1921