Displaying consecutive alerts from plugins should not hang the browser. The plugins display alerts via the NPN_Evaluate API. The browser signals an event handle to ensure that the plugin starts peeking for messages while waiting for the NPN_Evaluate call to return. When the dialog is dismissed by the user, windows does send some messages to the plugin window underneath, like activation messages, etc. These don't get dispatched as the event is reset when the dialog is dismissed, i.e. much before the window is actually destroyed.
The fix is to reset the event handle after the window is actually destroyed. To achieve this I added an OnClose virtual function to the DialogDelegate interface, which is overridden by the JavascriptMessageBoxDialog class which eventually ensures that the event is reset.

This fixes http://code.google.com/p/chromium/issues/detail?id=10799

I updated the AlertInWindowMessage npapi test to display two alerts instead of one.

Bug=10799

Review URL: http://codereview.chromium.org/113464

git-svn-id: http://src.chromium.org/svn/trunk/src@16172 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 92bbe3ef36b50c2ccb265ad0ed87fc191054db45