Fix test_shell so that we can pass LayoutTests/http/tests/navigation/relativeanchor-basic.html
If a navigation only consisted of scrolling to an anchor, then we would skip
the normal 'done' processing.  As a result, we would not update the 'content
state' of the NavigationEntry.  The code that dumps the back/forward list
inspects the navigation entries, and it would see a blank content state for
the last entry.

BUG=11924
R=dglazkov

Review URL: http://codereview.chromium.org/113530

git-svn-id: http://src.chromium.org/svn/trunk/src@16300 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e71487dd7c203ad83132c776051e397034794b0e