Factor out a BrowserFrame interface, and move the existing Windows-based code to a new class: BrowserFrameWin.
The major other change I had to make was the way that the TabStrip is set in
the BrowserRootView. Now there is a method on the frame that the view can
call when it creats the tabstrip that will set the tabstrip on the root view.
Review URL: http://codereview.chromium.org/113511

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16373 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b332599dc85c926caf7303653cd0cddd387ce042