Quick reimplementation of StatusBubbleGtk to not suck as much.
The TabContentsContainerGtk now uses a GtkFixed to store its
children, and the status bubble is now a child of that GtkFixed
so that it can be absolutely positioned on top of the rendered
data. Since it is no longer a GTK_WINDOW_POPUP, all the weird
stuff related to different window managers goes away.

http://crbug.com/11635

Review URL: http://codereview.chromium.org/113590

git-svn-id: http://src.chromium.org/svn/trunk/src@16433 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1c000db933af869a1236632bb87ac229d0c2b05f