Use av_rescale_q() for converting FFmpeg timestamps to base::TimeDelta.
Previously we were using integer math to convert to microseconds, but depending on the frame rate and packet size we could introduce enough error that could accumulate and introduce audio/video synchronization drift.  av_rescale_q() is a simple function but is designed to minimize error as much as possible.

Review URL: http://codereview.chromium.org/113598

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16453 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8b1dd688c0842bbd50023a0863811192d3905a2a