Fix a invalid read found by valgrind in process singleton linux.
We set up a listener on the IO thread, but the listener lived longer
than the IO thread.  Instead, register a listener to stop listening
when the IO thread goes away.  We have to register the destruction
listener on the thread that we're listening to, so there's an new
StartListening method dispatched from ProcessSingleton::Create.

Review URL: http://codereview.chromium.org/115605


git-svn-id: http://src.chromium.org/svn/trunk/src@16606 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6d224a2068e1caca549953bad1147c6f886b5549