Implement temporary First Run Dialog on OS X
We use a modal dialog with a single checkbox on OS X.

We use the OSX defaults system since we want something quick and reliable.  The
dialog is displayed at a very early stage in Chrome startup (Before any
subsystems start relying on the stats variable).  This means there are a few
quirks in displaying the UI.

A change was also needed to our event handling code since when the dialog is
shown we spin an event loop at a very early stage in the process lifetime.

Changed default value for stats to false and updated unit tests to reflect that.

Also some misc. minor cleanup.

BUG=11971,12046

Review URL: http://codereview.chromium.org/115608

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16669 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 89cd3137a89d8a7bdbeabf0be3dc6d79b46d679f