Remove the HistoryState property of WebRequest.
In the new WebKit API, it seems best if WebRequest is just a wrapper for
WebCore::ResourceRequest since in most contexts that's what we need it to
be.

The solution here is to introduce a LoadHistoryState method on WebFrame that
can be used to navigate to a session history item.

BUG=10038
TEST=covered by existing back/forward navigation tests.
R=brettw

Review URL: http://codereview.chromium.org/113758

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16747 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f281782435d2dbdfbb2d90f878c6c97d2713b3ec