Fix a bug in popup menus where we sent the wrong mouse events coordinates to the renderer.
gtk_grab_add() causes us to receive all events.  That means we receive
mouse events for other widgets with the coordinates relative to the other
widget.  We actually the coordinates relative to the popup.

This caused jankiness with the popup on news.google.com.  Since the popup happened to be on
the left side of the page, we sent event coordinates that happened to land in
the popup, causing us to select stuff.

BUG=12025

Review URL: http://codereview.chromium.org/113750


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16791 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9a6bf68e7bad502f0e0ba95b21722b6d2eb08696