Make sure that seekable streams are removed from the list of resource clients maintained by WebPluginProxy.
The PluginStreamUrl object ensures that it notifies the webplugin object that it is being deleted in its destructor. 

The other change is to remove the call to CancelRequest in WebPluginDelegateImpl::CreateResourceClient for
seekable streams, as this call does not do anything for manual streams created with a resource id of -1.
In any case for byte range requests, receiving a new response does not mean that earlier requests should
be cancelled.

This fixes bug http://code.google.com/p/chromium/issues/detail?id=12445

BUG=12445


Review URL: http://codereview.chromium.org/113772

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16828 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8831d1533b25c2354337cebe1a7191d6d1843db1