First phase of Mac Keychain integration. For the overall plan, see the design doc linked from the bug. This implements the ability to read existing Keychain passwords, but no storage/updating yet, and no adding of our own metadata as passwords are used; all that will be done in follow-up patches.
This also includes a Keychain wrapper for the functionality necessary so far, and a mock Keychain sufficient to unit-test essentially all of the code.

This patch deliberately excludes the step of instantiating a PasswordStoreMac and hooking it up to the profile, because a bug in autocomplete itself prevents passwords we load from actually being used, and we don't want to trigger Keychain UI for passwords that will be ignored. So this won't be used in builds yet, but it will be unit tested.

BUG=11745
TEST=none (no user-visible effect yet)
Review URL: http://codereview.chromium.org/115658

git-svn-id: http://src.chromium.org/svn/trunk/src@16897 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: dd2b4a444c8f0ea9cd2c066def2694f9d54b3a5c