Linux: Breakpad updates.
The crash server requires an MDRawSystemInfo stream in order to process the
file, although it actually only needs a single value in it.

Additionally, it needs a PDB7 structure for each module, so we remove those
modules which probably aren't shared libraries and XOR the first page of data
to get a signature for them. The previous code opened the file and MD5 summed
the text segment. However, we can't know if the shared library is still linked
on disk, nor if the whole text segment is mapped in memory. Thus, we only
assume that a single page is mapped.

dump_syms will need to be forked and updated to match the signature scheme used
in this patch.

http://codereview.chromium.org/113875


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17030 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d3050798248e23afff3fbf51954797c2981b0dae