Fix 12729: Find box doesn't unregister Esc accelerator.
This regressed recently during porting. When
comparing to see who is registered we need to 
be comparing against a class that implements
FocusChangeListener (and registers as such
with AddFocusChangeListener). In this case,
host_ does not implement that interface and 
doesn't register as a listener, so it will 
never be registered. We should be comparing
using 'this', not 'host_'.

I've fixed this and added a browser_test to 
catch this in the future.

BUG=12729
TEST=Open Find, press Esc, press Esc again 
and the Find box should not briefly appear
and hide again. Also, make sure Esc works 
to cancel navigations.

Review URL: http://codereview.chromium.org/115832

git-svn-id: http://src.chromium.org/svn/trunk/src@17032 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5ca7658a07930a7af87f5b3f9a7310e2132d1556