Fix layout test failures.
1- We need to be careful when converting from a null WebURL to a GURL
since std::string(NULL, 0) crashes.

2- It turns out that in some layout tests, willSendRequest sets the request to
null to indicate that we should not follow the redirect.  In the few cases I
debugged, this was happening because we were redirecting from "localhost" to
"127.0.0.1".  It seems like we probably need to change the hostname used to
load HTTP based layout tests to match what the tests expect.  For now, I just
commented out the assertion since it was something that I had just newly added.

BUG=none
TEST=covered by layout tests
TBR=dglazkov

Review URL: http://codereview.chromium.org/115970

git-svn-id: http://src.chromium.org/svn/trunk/src@17291 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e75a5aadbade308ef4fde7a9e278fe54389c00d8