Linux: stop libevent from pre-allocating huge arrays.
Currently, libevent allocates two structures of size RLIMIT_NOFILE
(which is 8192 on Ubuntu) and so these end up being several hundred kB
of memory for each MessageLoopIO.

This patch changes the default size to 32. If the thread watches a
file descriptor greater than 32, the arrays will be grown as needed,
doubling each time.

Also fixes a typo in the upstream code. A bug report for this has been sent
upstream.

http://codereview.chromium.org/118126
BUG=11999


git-svn-id: http://src.chromium.org/svn/trunk/src@17436 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f25eb4d7edf0a544038f96b3979ff10a9fca984b