Python script for auto-generating stub/def files for a list of signatures.
This python script takes a list of C-style signatures and generates either a windows def file, or a C++ stub file that allows for binding of functions via dlopen/dlsym.  This is used to support delay loading in windows and simulate it on posix systems.

Review URL: http://codereview.chromium.org/113803

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17534 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e5dd081e74432476d1c09c287838ac64fbee055a