Fix: Alt+Tab not sending FocusWillChange events.
This causes a DCHECK in FindBarWin if the Find box is open when the user switches to another app with Alt+Tab.

The Focus manager now calls ClearFocus everywhere it used to set focused_view_ = NULL, to make sure the event reaches all listeners for FocusWillChange.

BUG=13350
TEST=Open Chrome, press Ctrl+F then Alt+Tab. Swith focus back to Chrome and press Esc multiple times and you should not see the Find box flash briefly.

Review URL: http://codereview.chromium.org/118267

git-svn-id: http://src.chromium.org/svn/trunk/src@17678 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e531416c5b9e81eaff31e2a50b2954b2b5f08592