Fix an issue where themes would sporadically fail to install.
Trying to send decoded images over IPC didn't work too well.  Instead, we'll
write them to a file and have the browser slurp them in from there.  My first
instinct was to use SharedMemory, but that would require us to impose a limit
on the size of the decoded image data.

Also made sure that the undecoded images are deleted when we install.

BUG=13455
TEST=Try the repro steps in bug 13455 several times and make sure it works
every time.
Review URL: http://codereview.chromium.org/119255

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17797 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1f9e062770ecd2fa4d7f548a8118556cabb8de61