Reapply caching of the document JavaScript wrapper on the global object.
Because initialization of a context can now instantiate a document
wrapper, we have to be careful if the context initialization happens
when attempting to create that very same document wrapper.  We
therefore make sure that the context is initialized before checking if
a node already has a wrapper.
Review URL: http://codereview.chromium.org/118430

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17939 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5aa743a54d916478128f675bc3407c4422190807