Update DEPS to pull in upstream changes that remove RGBA32Buffer::bitmap(), and rewrite local tests to compile.  Also pulls in a change that upstreamed V8CustomBinding, since that got bracketed between my upstream changes; this includes the Chromium side of that too.
This change exposed a deficiency in the XBM testing where we were erroneously using a zero-length set of image data to calculate the MD5 sum (because we were using the buffer rect instead of the image size, and the XBM decoder wasn't correctly setting the rect).  Fixed the test to use the image size (which is more correct for this application), will fix the XBM decoder upstream to set the rect correctly.

This in turn required generating new MD5 sums, which required patching the generation code (which is normally #ifdefed away) to also work with the past couple years of changes.

This commit will need to land at the same time as one that updates the expected sums; during the period where either commit is in place without the other, the XBM decoding tests will fail.

BUG=13633
TEST=Covered by unittests
Review URL: http://codereview.chromium.org/121001

git-svn-id: http://src.chromium.org/svn/trunk/src@18131 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0a3cd451ed3ebb891825f4e5ba11f23f6b449457