Http Cache: First pass of byte-range requests support.
This is the first pass to implement support for range
requests and the asociated sparse cache entry.

It is disabled by default, and requires ENABLE_RANGE_SUPPORT
to be defined in order to activate the code: all the code
is compiled in, but Start() bypasses the cache for range
requests, and OnNetworkInfoAvailable ignores 206.

Big parts are still not implemented. Most notably we are
not modifying the response headers that we send to the
user, so the content-range and content-length info
are not correct.

BUG=12258
TEST=Unit tests.

Review URL: http://codereview.chromium.org/118345

git-svn-id: http://src.chromium.org/svn/trunk/src@18289 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d633c89da8bcdddb488369ce4f1edcbf8a650ed5