Copy login-related section of WebDatabase into a new class for use in PasswordStore.
The platform-specific inheritence here is because the current Windows implementation of PasswordStore is a temporary shim, and will eventually need to use its own database instead; since we'll need almost all the fields on the Mac for the bag-on-the-side anyway, it makes sense to share as much of this implementation as possible.

The code and unit tests are pretty much unchanged from WebDatabase (except for the inheritence stuff); the duplication should be temporary as the WebDatabase copy will be removed once Windows switches (and migrates their passwords).

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/119392

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@18330 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5c50865947a190023f441ffabb9adbd160dc986a