Copy login-related section of WebDatabase into a new class for use in PasswordStore.
The platform-specific inheritence here is because the current Windows implementation of PasswordStore is a temporary shim, and will eventually need to use its own database instead; since we'll need almost all the fields on the Mac for the bag-on-the-side anyway, it makes sense to share as much of this implementation as possible.

The code and unit tests are pretty much unchanged from WebDatabase (except for the inheritence stuff); the duplication should be temporary as the WebDatabase copy will be removed once Windows switches (and migrates their passwords).

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/119392

git-svn-id: http://src.chromium.org/svn/trunk/src@18330 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c2192ab93a6ef8baa50fe53326ff5654b6d4ebb2