Make findbar paint themed image.
Also, fixed drawing of background images that are not 1 pixel wide. (The code for this is somewhat speculative as there is no way of testing it at the moment, but for now it should still at least look correct.)

If you open the findbar while the bookmark bar is opened and then close the bookmark bar, the findbar painting is a little messed up in that it still thinks it's farther away from the tabstrip than it really is. The solution (unfortunately) is to continuously repaint the findbar as it is closing. Since this is glitch is hardly noticeable, I'm putting off fixing it.

TEST=findbar should look more integrated with bookmark bar, toolbar.
BUG=none

Review URL: http://codereview.chromium.org/125084

git-svn-id: http://src.chromium.org/svn/trunk/src@18339 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e0cd384d34e6b733615ca94c33b2e340ee2812e7