Properly cleanup state after synchronous DNS resolution
Pre-resoultion system was not cleaning up its state when the resolver
returned asynchronously.  This regression probably landed when we shifted
to using an internal resolver class, rather than running on worker threads.
The impact was made detectable by the recent introudction of a micro-cache,
which often returns with a synchronous result.

BUG=14212

r=eroman
Review URL: http://codereview.chromium.org/125209

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@18537 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0f144d91531e9ba59357bd88daf21abd3231263a