Adds kind-of-live thumbnail generation for a potential tab switcher.
This listens to tab events and tries to keep thumbnails ready to go. See
thumbnail_generator.cc for a more detailed design.

This adds a painting observer to the RenderWidgetHost to enable this new
behavior, as well as a notification to allow the thumbnail generator to hook
its observer in. There is also a new notification that a backing store has been
disabled, which required making the backing stores know about their owning
widget hosts.

This component is currently disabled. We just need to uncomment the member in
Profile and it will start to work.

Original review: http://codereview.chromium.org/118420

Review URL: http://codereview.chromium.org/126101

git-svn-id: http://src.chromium.org/svn/trunk/src@18540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a93165ca42bb2e497b498e42f23841d08ada6ef8