Re-land final subset of r18520.
The original code review that this belongs to is: <http://codereview.chromium.org/125171>. (BUG=14138)

The original was backed out because of a valgrind linux error (BUG=14218), so I relanded it in smaller chunks to see at which point problem happens.

This differs from the original, in that I have changed the unit-test to no longer use ScopedMapper + WaitingMapper, which I believe was the source of a race.

Review URL: http://codereview.chromium.org/125232

git-svn-id: http://src.chromium.org/svn/trunk/src@18587 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 137255c4d7eb212e24d553e4740bf42919ef1245