Linux: dump_app_syms would fail if the output didn't exist.
This script was trying to only rebuild the symbols when the binary
file was more recent than the symbols file. However, -nt fails when
the right hand side doesn't exist. Therefore it wouldn't actually dump
symbols and the sed script would fail later on.

This adds a test for the non-existance of the output file first.

TBR=mmoss


git-svn-id: http://src.chromium.org/svn/trunk/src@18659 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8f2db6e51e317e82a0a51c68b19f4b185b35a2bc