Linux: dump_app_syms would fail if the output didn't exist.
This script was trying to only rebuild the symbols when the binary
file was more recent than the symbols file. However, -nt fails when
the right hand side doesn't exist. Therefore it wouldn't actually dump
symbols and the sed script would fail later on.

This adds a test for the non-existance of the output file first.

TBR=mmoss


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@18659 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1756b401ac1ff4c103f2ed571eff4864448fa2fc