Enable SDCH on mac/linux
Although it compiles in all builds etc., someone left the enabling code
ifdef'ed in browser main. This should enable SDCH on mac/linux.

To test, visit a site that supports SDCH (twice... the first time
you'll get the dictionary in the background), and then do a
view-cache:
in front of the resulting URL, and check to see that you're getting
Content-Encoding: sdch,gzip
If it is not sdch encoded, then with some tiny (1%) probability, 
you're in a control group of a current experiment... and you can
try starting your browser again... and doing another two visits.

To better track the issue if you think you are in the control group,
you can also visit
about:histograms
and you should see that a dictionary was downloaded, and then
either some content was decoded, with timings in Sdch3.Experiment_Decode
or was not run in SDCH and was timed in Sdch3.Experiment_Holdback.

BUG=14516

r=willchan,pinkerton
Review URL: http://codereview.chromium.org/131025

git-svn-id: http://src.chromium.org/svn/trunk/src@18753 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a67e5864e4cfd742e2986217cbebaa761e1d11e2