Fix crash in LocationBarView when accessing TabContents.
The LocationBarView asks its delegate (the ToolbarView) for a pointer to TabContents. The problem is, the ToolbarView is keeping an internal pointer to TabContents, and that pointer is not getting cleared when the last tab goes away. This can lead to shutdown crashes, which is what I think is happening in the crash in bug 14601. 

This is a speculative fix because we can't reproduce, but according to the dis-assembly we access a corrupt STL struct through a non-null TabContents pointer, which makes sense if the TabContents pointer is not getting cleared (or if the STL struct is being modified through multiple threads, which I have verified it is not).

BUG=14601
TEST=Cannot reproduce, we'll have to watch the crash reports as they come in to see if the crash is fixed.

Review URL: http://codereview.chromium.org/132043

git-svn-id: http://src.chromium.org/svn/trunk/src@18775 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1964726946de195ec41fe57e96441092946c2c85