linux: make extensions service test pass in Debug.
It only passes on the build bots because they still have old directories, I
think?  AbsolutePath() on POSIX will fail if the path doesn't exist.  We let
the program march on in that case and the normal error handling (that would
handle the non-existent path) make the tests continue to pass.

Review URL: http://codereview.chromium.org/132054

git-svn-id: http://src.chromium.org/svn/trunk/src@18834 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1322b249d9a4dbe093cdcd5b5f8d37f30cfd5559