This patch adds Cocoa editing selectors to the RenderViewWidgetHostMac class.
This makes us a better citizen in OSX text-editing land and is a first step in getting our text entry story to work more natively on OSX.

The selectors are added at runtime to the Cocoa class and invoking them causes the appropriate WebKit core command to be invoked on the currently focused WebFrame.

As a side-effect bug 10862 is fixed.

BUG=10862
TEST=Open a new tab and type some text into an input element, change to a Hebrew keyboard layout and hit command-a, all text should be selected.

Review URL: http://codereview.chromium.org/114070

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@18853 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2544ac8acbcefb5dc6348b0f8c675aac8574c521