Add the ability to resize the window when over the custom frame.
I also switched from 3px borders to 4px borders because that's what they are
on windows.  On linux, the borders look bigger because our webcontent area
doesn't drop a shadow.

For this to work for resizing on top, I had to remove the event box from
the browser titlebar and move the mouse event handling to the window.

BUG=14645

Review URL: http://codereview.chromium.org/140026


git-svn-id: http://src.chromium.org/svn/trunk/src@18927 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 736bd59b42edcc332341d6c57abdc65b274b3f0d