Linux: fix crash reporting for zygote model
Crash reporting broke on Linux when we enabled the zygote model a
couple of weeks ago.

We can't just add "zygote" to the check for the process type because
the crash signal fd is set at the same time and that will change when
a renderer is forked from the zygote.

This fixes it for now, but it will need to be redone when we enable
sandboxing.

http://codereview.chromium.org/147004
BUG=14969


git-svn-id: http://src.chromium.org/svn/trunk/src@19035 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9a1f3b21409eae34aa8c49fc9cd156f4b1c16724