Start to disable bzip2 support.  First step: Stop advertising support.
Although it would be nice to just yank it all, it is good to do
so slowly so that any cached data that might required bzip2 decoding
will still decode and render.  Maybe I'm being paranoid... but this seems
like a nice first small step.

I also disabled the tests that depend on this modified code, and linted
the unittest file a bit.

BUG=14801

TEST= Be sure you can still view http://ukcp09.defra.gov.uk/content/view/16/6/index.html which is (currently) bzip encoded.

r=hunar
Review URL: http://codereview.chromium.org/147058

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19173 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4b1e2cd9ed3e3b498f4e8180041eea70d0e4b816