Replace std:;wstring with std::string and string16 in locale-name related APIs.
1. Change the locale param to be std::string because they're always ASCII and change call-sites accordingly.
2. Add GetStringFUTF16 to l10n_util. On Windows, they're inline helpers calling the correspondingGetStringF returning wstring while on Mac/Linux, they just return the result of |string16 GetStringF|without converting to wstring.

This is part 1 of the fix for issue 8647. Some of newly introduced conversions are temporary and will be removed later (e.g. ASCIIToWide applied to the result of GetApplicationLocale in a few places). 

Note : this CL will be landed after http://codereview.chromium.org/147038 is landed. 
BUG=8647 (http://crbug.com/8647)
TEST=Pass l10n_util_unittest and other unit tests

Review URL: http://codereview.chromium.org/126223

git-svn-id: http://src.chromium.org/svn/trunk/src@19183 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 88e5c9e9770ce741eb2ca86a1c8c164f6d3f3287