implement WebSandboxSupport stub to prevent a null ptr deref
If FontChromiumWin's Font::drawGlyphs() fails for some reason, we
may attempt to reload the font using ChromiumBridge::ensureFontLoaded().
This requires a sandbox routine, SandboxSupport::ensureFontLoaded() to be
implemented, and in test_shell, it wasn't, causing a null ptr deref and
weird things to happen.

Note that there is no easy way to provoke this codepath once the other
bugs causing the crash in #10977 are fixed.

R=darin
BUG=http://crbug.org/10977
TEST=none


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19200 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ed2bae7ea64b856c43e058ef71ec87fcf082309e