Linux: add linux-gate.so to the minidump mappings.
Add linux-gate.so to the minidump mappings list by checking for the
load address in the aux vector and adding a filename of linux-gate.so
when we encounter that load address. This is necessary for crash
reporting to walk the stack correctly when a crash happens while in a
system call.

http://codereview.chromium.org/147032

Committed for Neal Sidhwaney.


git-svn-id: http://src.chromium.org/svn/trunk/src@19214 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7d04b8bad2c2cd63fbecc34fe1028fbb1715cd2a