Linux: plumb fontconfig call out to the sandbox host.
This is hopefully the last step before our renderers can run cleanly
in a chroot.

WebKit needs to be able to ask for the correct font to use in the case
that the current font doesn't include glyphs for certain code points.
Currently we make a fontconfig call in our WebKit port to handle this.

This patch changes this so that the call is sent our via
ChromiumBridge. Since we are at ChromiumBridge, we could make a sync
IPC to the browser. However, fontconfig is a single threaded library
and we are already using it on the UI thread in the browser, so the
sync IPC would have to terminate on the UI thread. Even if this
doesn't deadlock, it causes huge spikes in latency.

So, instead, we send the IPC to the sandbox host process which is
already setup to handle fontconfig requests from Skia. See:
  http://code.google.com/p/chromium/wiki/LinuxSandboxIPC


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19287 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 80bfb745d06c79df2d1149a564e10ece1485a9a9