Linux: fix font fallback for charactors missing from the primary font.
We are using the sandbox host for font fallback now via the WebKit
API. However, a patch landed in WebKit which added an assertion about
being on the correct thread. Since the sandbox host was using
WebString::fromUTF8, we hit the assertion and died.

So, we have to implement a WebKitClient for the sandbox host. We
cannot just pass in NULL because there's an assertion that it's
non-NULL. Also, we cannot pass in a bogus pointer because it actually
ends up getting called back.


git-svn-id: http://src.chromium.org/svn/trunk/src@19324 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5ba873fa625704c0a557a2a82c6f4d3ea81beb55