Fix download shelf z ordering during close.
Pretty much the same as what we do in the find bar. I think the flickering thing in the bug desc. was another result of <http://crbug.com/14786>, as it was only present in 3.0.190.2 and not ToT, and toggling the bookmark bar fixed it.

http://crbug.com/15427
TEST=see bug

Review URL: http://codereview.chromium.org/149054

git-svn-id: http://src.chromium.org/svn/trunk/src@19401 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b117520ebff82cd6410358902a3fda68764b2a6d