Rework a loop in the native menu code to iterate over the vector of items we have, rather than the actual system menu.
I believe this may fix a crash where the code indexes out of bounds in the vector. The trouble is the assumption that the native menu index matches the index into the vector. If someone has installed a windows addon this may not be true. It's safer to restrict the iteration to the bounds of the vector and index into the menu based on that.

http://crbug.com/14600
TEST=none

Review URL: http://codereview.chromium.org/150061

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19558 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2bc0ffd998d2ad996123f2d446a77a82d544e6c5