linux: only create browser-side plugin container after plugin requests it
On Windows, windowed plugins are parented in a two-stage process: initially,
the plugin is parented to the render view, then later it's reparented to
a special per-plugin window.

On Linux, plugin embedding always needs a special per-plugin window before
it can be initialized.  So there's no way to do the two-stage initialization
process, and we only want to construct the browser-side plugin container for
windowed plugins after the plugin requests it.

Plumbing this through the WebPlugin interface will also allow us to side-route
this request for window creation out to the browser process in the
multi-process case.

BUG=15421
TEST=plugins still work in test_shell

Review URL: http://codereview.chromium.org/150034

git-svn-id: http://src.chromium.org/svn/trunk/src@19560 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0d4ae4df89be96057a4cc405c4fd735202bab276