linux: only create browser-side plugin container after plugin requests it
On Windows, windowed plugins are parented in a two-stage process: initially,
the plugin is parented to the render view, then later it's reparented to
a special per-plugin window.

On Linux, plugin embedding always needs a special per-plugin window before
it can be initialized.  So there's no way to do the two-stage initialization
process, and we only want to construct the browser-side plugin container for
windowed plugins after the plugin requests it.

Plumbing this through the WebPlugin interface will also allow us to side-route
this request for window creation out to the browser process in the
multi-process case.

BUG=15421
TEST=plugins still work in test_shell

Review URL: http://codereview.chromium.org/150034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19560 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d2697829eb98ca20ce799e9ccc692c33c8ffee0c