Update the browser window's shape on state change.
When launching the browser we restore the maximized state. However, the maximize call is asynchronous, so we don't actually observe the change to |state_| until later. So we need to update the shape on that later state-changed signal.

Usually if the user maximizes the window, it will change sizes. To avoid setting the window's shape more often than necessary, cache its current state (custom shape or no custom shape).

BUG=none
TEST=maximize the window, close chrome, re open chrome. It should not have a custom shape.

Review URL: http://codereview.chromium.org/150081

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19625 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d5b98cb0d0d8aef66df8d86ffbfbac0ff78f566d