linux: make layout test plugin build and run in test_shell
I used the upstream implementation as a guide for the changes I made,
so style faults are theirs.

Also includes updates to test_expectations for the new tests we pass,
and one new baseline for a test that had an output image that we match
Windows on aside from antialiasing.

BUG=11516

Review URL: http://codereview.chromium.org/150144

git-svn-id: http://src.chromium.org/svn/trunk/src@19698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 13ff5f21223bff68330e99e0f19a8ee7599181a4