Another batch of performance improvements for NNTP.
Hard code the DOM and the layout for the default case so that we do not
need to do any js layout at startup.

Be more careful when calling layout. I added a dirty bit so that we
don't layout too often.

CSS cleanup/optimization to make selectors faster.

TEST=Run the perf test

BUG=13362

Review URL: http://codereview.chromium.org/149127

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19877 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 23d719d0f61aca2c2fb163105bd26abb868f0c4c