linux: fix a layout test by having our test plugin provide two mimetypes
The layout test LayoutTests/fast/js/navigator-mimeTypes-length.html asserts
that the number of mimetypes handled by plugins should be greater than the
number of plugins.  This isn't true if the layout test plugin is the only
plugin and it only handles one mimetype, so we make it handle an additional
fake mimetype.

Review URL: http://codereview.chromium.org/149146

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19881 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c20756b95d30ccffb843f17135e9027a1af81117