linux: fix a layout test by having our test plugin provide two mimetypes
The layout test LayoutTests/fast/js/navigator-mimeTypes-length.html asserts
that the number of mimetypes handled by plugins should be greater than the
number of plugins.  This isn't true if the layout test plugin is the only
plugin and it only handles one mimetype, so we make it handle an additional
fake mimetype.

Review URL: http://codereview.chromium.org/149146

git-svn-id: http://src.chromium.org/svn/trunk/src@19881 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 84da8433987acb421aa4ab0dd6ef4e41cd30361b