Make action buttons listen to the enter key.
Hide menus when the window is blurred or the user presses Alt or Meta.

Only show tootlip for the title of the thumbnail.

Allow mouse over on the window menu.

Made the windows opaque but with a hsla background color to make it more readable.

Quote all background image URLs so that we can display icons for URLs with spaces in them.

Call preventDefault when using the keyboard to navigate the options menu so that the page does not scroll.

BUG=15268, 15503, 15715, 15769, 15411, 15458

TEST=See bug descriptions

Review URL: http://codereview.chromium.org/149163

git-svn-id: http://src.chromium.org/svn/trunk/src@19884 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d6800bef7e381967cf6870ceb254cc6fc34b0729