Fixing a crash in RenderViewGone.
If I start Chrome and shut it down quickly RenderViewHost
receives an IPC ViewHostMsg_RenderViewGone. We respond in
ExtensionHost::RenderViewGone by putting up an info bar on
the selected tab contents saying that an extension
crashed. 

Only problem is that ExtensionHost::GetBrowser() can
return NULL during startup and shutdown, as documented
in the function and so we should not try to put up the
Infobar if the browser is going away.

BUG=None
TEST=Open Chrome with extensions enabled and an extension
loaded and shut Chrome down as soon as it launches. It
should not crash. Might take a few trys to repro.

Review URL: http://codereview.chromium.org/149205

git-svn-id: http://src.chromium.org/svn/trunk/src@19986 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 85e69e3a257e570ad1aeb1b09b787f67ccb1e9ca