Adding a non-templatized version of QI for times when the IID and type aren't associated.
Also fixing the ScopedComPtrVector unit test.
The problem with the test was that it expected AddRef to be called twice and Release() once when pushing a pointer to the vector.  However that only happens when a push causes a reallocation.  When there's no reallocation, there's only a single call to AddRef and no call to Release, causing the test to report failure.  The fix is to simply use std::list instead and fix the expected values.

TEST=Run *ScopedComPtr* unit tests.
BUG=none
Review URL: http://codereview.chromium.org/149345

git-svn-id: http://src.chromium.org/svn/trunk/src@20183 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a9f37b2ed54dfbce8e8c8f30bed0bf0732b8aaf4