linux: implement the sad plugin
This required moving the sad plugin resource out of the Windows-specific code.

Note that it doesn't display right yet, since the place to display is covered
by the GtkSocket that used to host the plugin.  But that is a separate issue.
(I had it draw to the side to verify the image was correctly getting loaded.)

BUG=16158

Review URL: http://codereview.chromium.org/155262

git-svn-id: http://src.chromium.org/svn/trunk/src@20264 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: cf5501d2a7202deaa3f35227ecbd315877f32a9d