linux: implement the sad plugin
This required moving the sad plugin resource out of the Windows-specific code.

Note that it doesn't display right yet, since the place to display is covered
by the GtkSocket that used to host the plugin.  But that is a separate issue.
(I had it draw to the side to verify the image was correctly getting loaded.)

BUG=16158

Review URL: http://codereview.chromium.org/155262

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20264 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 894c0d026cba268fb4fda50c36b52c7809e7a197