Add a rudamentary feed preview to the RSS extension. It doesn't handle inline HTML in the item description (it just dumps it as text) and the feed needs to be valid XML for it to show any preview, but it is better than nothing. We can easily change it to display the HTML but we want to (at some point) try to use a separate origin so that we can render the HTML code from untrusted sources safely.
Also fix a bug in the image tracker. It should not try 
to communicate with the view if the view has gone away
(which was the whole point of the image tracker...)

BUG=None
TEST=Install the extension, browse to a page with a feed
and click onthe rss icon in the Omnibox. An interstitial
page should appear with a preview of the feed.
Review URL: http://codereview.chromium.org/155180

git-svn-id: http://src.chromium.org/svn/trunk/src@20316 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ad674e154216bfccf58d6be0c74059d973fd69bb