linux/mac: Fix return status of OpenChannelToPlugin message in case of failure
Currently on linux if the plugin process dies, the PluginProcessHost is not destroyed.
A side effect of that is that if a renderer tries to create a new plugin, the OpenChannelToPlugin message returns success with an empty handle, which is interpreted as the host channel, to which the renderer tries to connect to a second time, causing DCHECKs and havok in the browser process.
This CL fixes that particular problem, returning failure if the browser can't connect to the plugin process.

Review URL: http://codereview.chromium.org/155313


git-svn-id: http://src.chromium.org/svn/trunk/src@20337 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7f5c88839433f21c57bc049edae3d7788a370d5e