Linux: use a temp directory for the chroot.
Ubuntu systems (at least) wipe /var/run at boot time, which is
deleting our sandbox directory. Instead, we have the SUID helper
create a temp directory in /tmp, unlink it and use that for the chroot
directory. A file descriptor is passed to the zygote process for it to
fchdir into.

(Thanks to fta for discussions on this.)

BUG=16363


git-svn-id: http://src.chromium.org/svn/trunk/src@20388 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5e4777ca1d72b7aa4ce8aac24e3926a56f8aed6e