Linux: use a temp directory for the chroot.
Ubuntu systems (at least) wipe /var/run at boot time, which is
deleting our sandbox directory. Instead, we have the SUID helper
create a temp directory in /tmp, unlink it and use that for the chroot
directory. A file descriptor is passed to the zygote process for it to
fchdir into.

(Thanks to fta for discussions on this.)

BUG=16363


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20388 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2c0751815fc22804526d85d1f1afd49fdca982d3