A little bit of bookmark bar cleanup, a little bit of padding improvement.
The spacing between the toolbar buttons now matches windows pixel for pixel (tested empirically).

The bookmark bar item spacing is now a little better, but we need some hackery to improve it to the state of windows (we need to annex about 5 pixels from the toolbar when the bookmark bar is open).

The code for packing bookmark bar buttons is now all in one place.

Functional change: There will be no other bookmarks button until the model is loaded. If you care I can fix this, it just doesn't seem like a very important matter since it's unusable anyway when the model isn't loaded.o

I also removed a log warning from chrome_dll_main because with this change it started spamming a lot. We're already living with this warning, and have a bug filed for it (they fixed the root problem in newer versions of gtk so I think this bug will just never get resolved), so I don't think this is a big deal.

BUG=15870
TEST=things look marginally better, theming still works properly
Review URL: http://codereview.chromium.org/155342

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20389 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e11cb53f150bb0ec689310c9862762457c1cbac1