Ensure proper paths when saving pages with no title.
When saving a page with no title, such as a text file, attempt to
use a sane value for the save name by retrieving the last component
of the URL (if one exists).

This prevents the case where a page http://www.foo.com/a/path/name.txt
is saved as file "http www" with extension type "foo.com a path name.txt".

This change also makes the SavePackage code a little more flexible
and simpler to test.

BUG=none
TEST=Covered by unittest.

Review URL: http://codereview.chromium.org/155266

git-svn-id: http://src.chromium.org/svn/trunk/src@20432 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7a1031fe8cb5e27e1bd01c8fec7ec635c45c7780