Document the O3D binary formats.
I just wrote up some small docs to kind of get this out
of the way. If Josie wants to write a more formal doc
to put on the website we can do that but at least with
this it's documented.

Going through it I have a minor concern. The Buffer
format is fine because it spells out exactly how much
data it expects.

The Skin and Curve formats do not so if pass in a
valid offset but an invalid length when you call
Cuvre::Set(myRawData, validOffset, invalidLength)
you get undefined results and possibly no errors 
depedning on what data it runs into.

This may or may not matter as it's unlikely the
user will get very far with an invalid length. The
issue is just more that it's possible.


Review URL: http://codereview.chromium.org/155246

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20457 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d8123b02244943a1616ed5a5e1fd1cb39623fa17