GTK: Force a size on download items and elide the dangerous download filename.
This is a temporary hack that might not work in all locales and definitely won't work if we ever let users change font sizes. GtkLabels are just plain broken. Pango is also broken because you can get height-for-width for a layout but not width-for-height.

Thus we should switch to using skia to draw download items fonts, but this solution will work for now.

BUG=16389
TEST=download the Dansk version of open office for windows

Review URL: http://codereview.chromium.org/155395

git-svn-id: http://src.chromium.org/svn/trunk/src@20572 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f8ba09d97a809e7f2aaf6d201bbb3f42e4fd45aa