Chrome-side of fixes for content-script messaging.
This change adds registration of content scripts, parented to a frame's
context.  When a frame's context goes away, we unregister it and any content
script contexts for it.

There's a corresponding webkit change that lets us know when a content script
context is created.  Filed upstream as
https://bugs.webkit.org/show_bug.cgi?id=27104.

BUG=16228
TEST=install an extension with a content script that communicates with a parent
process.  Messages should be sendable both ways.

Review URL: http://codereview.chromium.org/155309

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20677 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 05c413067347ec51daa9c5e33767962b8925f8a2