linux: destroy browser-side plugin windows on plugin crash
Because the plugin windows must be destroyed by the UI thread,
we store the map of plugin PID -> plugin windows there,
on the RenderWidgetHostView.

This differs slightly from Windows, where the map is stored
on the PluginProcessHost, but is otherwise pretty similar.

TEST=crash a windowed plugin; you should see a sad plugin
icon in the place of the plugin

Review URL: http://codereview.chromium.org/155517

git-svn-id: http://src.chromium.org/svn/trunk/src@20683 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 020ef046446754807410734e5c4535ae60116db3