Make more things clickable in the bookmark bar menus:
- folder within a folder
- "empty" menu item

Note that it can be hard to click on a folder within a folder since hovering over the folder pops up a submenu that grabs focus, so you end up needing to right click twice. It's a little odd but the hackery required to fix it doesn't seem worth it.

BUG=15886
TEST=right click on an empty menu item

Review URL: http://codereview.chromium.org/155586

git-svn-id: http://src.chromium.org/svn/trunk/src@20785 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 39ff4a410a22e4653628310b27f2b92eede36ad5