Linux: add use_system_ffmpeg GYP variable.
In r20548, we switched from using a shell script to copy the ffmpeg
binaries to the output directory, to using the GYP 'copies' stanza.
When built using Scons, any copies for which the source file was
missing began to fail.

Since Ubuntu policy (and size considerations) prohibit the shipping of
the binaries in the source packages of chromium-browser, so add a GYP
variable which can be set to inhibit this copying for cases where the
source files aren't included.

http://codereview.chromium.org/155791


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21093 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f884da5e7d5d6299c9273603509bf4a4390c8f2c