Make sure installer_util_string is always built before installer_util. It was triggering a bug in Incredibuild.
By making one of the dependent of install_util a dummy_executable,
and making this dummy_executable depends on installer_utils_strings,
the odds that it triggers a bug is a lot smaller. I tried 6 clobber
builds so far and none of them failed.

Review URL: http://codereview.chromium.org/155923

git-svn-id: http://src.chromium.org/svn/trunk/src@21283 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: aa7f4ca27efd781df41fa30ee2e3c09f9c89089e